requirements.txt
langchain>=0.0.267
openai>=1.3.0
selenium>=4.10.0
PyQt5>=5.15.9
flask>=2.3.3
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
nest_asyncio>=1.5.7
docker>=6.1.3
pydantic>=2.4.0
langchain-openai>=0.0.2
flask-cors>=4.0.0 